Move PWA Special Events DisplayCard into RCL for shared use.
Share the public event card UI and FormVM so Admin can preview current events as the PWA shows them via a toggle on Special Events.

Fixes #190
