+++ PWA/Features/Home/SpecialEventsCard.razor
@@ -1,5 +1,5 @@
-@using PWA.Features.SpecialEvents
 @using PWA.Features.SpecialEvents.Data
+@using RCL.Features.SpecialEvents
 
 @inject ISpecialEventsApiService SpecialEventsApi
 @inject ILogger<SpecialEventsCard> Logger
