+++ PWA/Features/SpecialEvents/Data/SpecialEventsApiService.cs
@@ -2,6 +2,7 @@ using System.Net.Http.Json;
 using System.Text.Json;
 using Microsoft.Extensions.Logging;
 using PWA.Features.SpecialEvents.Constants;
+using RCL.Features.SpecialEvents;
 
 namespace PWA.Features.SpecialEvents.Data;
