+++ PWA/Features/SpecialEvents/DisplayCardList.razor
@@ -1,4 +1,5 @@
 @using PWA.Features.SpecialEvents.Data
+@using RCL.Features.SpecialEvents
 
 @inject ISpecialEventsApiService SpecialEventsApi
 @inject ILogger<DisplayCardList> Logger
