--- PWA/Features/SpecialEvents/FormVM.cs
@@ -1,21 +1,0 @@
-﻿
-namespace PWA.Features.SpecialEvents;
-
-public class FormVM
-{
-	public int Id { get; set; }										// [Required]
-	public DateTime? ShowBeginDate { get; set; }	// [Required]
-	public DateTime? ShowEndDate { get; set; }		// [Required]
-	public DateTime EventDate { get; set; }				// [Required]
-	public int EventTypeId { get; set; }          // [Required]
-	public string? EventTypeDescr { get; set; }   // From EventType.Descr (API join)
-	public string? Title { get; set; }  
-
-	public string? SubTitle { get; set; }
-	public string? ImageUrl { get; set; }
-	public string? YouTubeId { get; set; }
-	public string? WebsiteUrl { get; set; }   // [DataType(DataType.Url)]
-	public string? WebsiteDescr { get; set; }
-	public string? Description { get; set; }  // ToDo: md?, probably going to be Component Body
-}
-
