+++ PWA/_Imports.razor
@@ -12,3 +12,4 @@
 @using PWA
 @using PWA.Layout
 @using RCL.Components
+@using RCL.Features.SpecialEvents
