+++ RCL/RCL.csproj
@@ -17,6 +17,7 @@
 		<PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="10.0.*" />
 		<PackageReference Include="Ardalis.SmartEnum" Version="8.2.0" />
 		<PackageReference Include="Azure.Storage.Blobs" Version="12.29.1" />
+		<PackageReference Include="Markdig" Version="1.3.2" />
 		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="10.0.*" />
 	</ItemGroup>
